+++ libgimp/gimpplugin.h
@@ -159,6 +159,10 @@ GList         * gimp_plug_in_get_temp_procedures    (GimpPlugIn    *plug_in);
 GimpProcedure * gimp_plug_in_get_temp_procedure     (GimpPlugIn    *plug_in,
                                                      const gchar   *name);
 
+void            gimp_plug_in_extension_enable       (GimpPlugIn    *plug_in);
+void            gimp_plug_in_extension_process      (GimpPlugIn    *plug_in,
+                                                     guint          timeout);
+
 
 G_END_DECLS
