+++ plug-ins/help/help.c
@@ -26,6 +26,8 @@
 
 #include <glib.h>
 
+#define GIMP_DISABLE_COMPAT_CRUFT
+
 #include "libgimp/gimp.h"
 
 #include "gimphelp.h"
@@ -206,7 +208,7 @@ help_run (GimpProcedure        *procedure,
       help_temp_proc_install (plug_in);
 
       gimp_procedure_extension_ready (procedure);
-      gimp_extension_enable ();
+      gimp_plug_in_extension_enable (plug_in);
 
       g_main_loop_run (main_loop);
